refactor: footer chunks into components
