+++ src/components/footer/Component.js
@@ -3,17 +3,14 @@ import React from 'react'
 /* eslint-disable no-unused-vars */
 import PropTypes from 'prop-types'
 
-import './styles.css'
 import BlankSpace from '../blankSpace/Component'
-import packageJson from '../../../package.json'
+import './styles.css'
+import Copyright from './copyrightComponent'
+import VersionInfo from './versionInfoComponent'
+import SourceCode from './sourceCodeComponent'
 
 /* eslint-disable no-unused-vars */
 const Footer = (props) => {
-  const curr_year = new Date().getFullYear()
-  const REACT_VERSION = React.version
-  const APP_VERSION = packageJson.version
-  const REDUX_VERSION = packageJson.dependencies.redux.replace(/[^0-9.]/g, '')
-  const WORKBOX_CORE = packageJson.dependencies['workbox-core'].replace(/[^0-9.]/g, '')
 
   return (
     <React.Fragment>
@@ -22,52 +19,14 @@ const Footer = (props) => {
           <div className="row mx-0 react-banner hide-small">
             <div className="col-12">
               <div className="text-center mr-3 mb-0">
-                © Copyright 2015 - {curr_year}. Progressive web app version {APP_VERSION}
-                <BlankSpace />
-                built with
-                <BlankSpace />
-                <img
-                  key="1"
-                  className="react-logo"
-                  src="/assets/images/react-logo-300x261.png"
-                  alt="ReactJS logo"
-                />
-                ReactJS {REACT_VERSION},<BlankSpace />
-                <img
-                  key="1"
-                  className="react-logo"
-                  src="/assets/images/redux-logo.svg"
-                  alt="ReactJS logo"
-                />
-                Redux {REDUX_VERSION} and
-                <BlankSpace />
-                <img
-                  key="1"
-                  className="react-logo"
-                  src="/assets/images/workbox-logo.svg"
-                  alt="ReactJS logo"
-                />
-                Workbox {WORKBOX_CORE}
-                <BlankSpace />
-                <a className="mx-1 learn-more-link" href="/reactjs" target="_self">
-                  Learn more
-                </a>
+                <Copyright /><BlankSpace /><VersionInfo />
               </div>
             </div>
           </div>
           <div className="row justify-content-center mb-0 hide-small">
             <div className="col-auto mb-0">
               <p className="mb-1">
-                Source code:
-                <BlankSpace />
-                <a
-                  className="edx-links"
-                  href="https://github.com/lpm0073/lawrencemcdaniel.com/"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  https://github.com/lpm0073/lawrencemcdaniel.com/
-                </a>
+                <SourceCode />
               </p>
             </div>
           </div>
